move bricks file to abstract folder
