+++ src/main/java/game/bricks/StoneBrick.java
@@ -1,6 +1,7 @@
 package game.bricks;
 
 import game.Constant;
+import game.abstraction.Bricks;
 import javafx.scene.image.Image;
 
 public class StoneBrick extends Bricks {
